ND events: source from CSV export instead of HTML events table
Web events scraper was going really slow for me, seemed to need to make a bunch of requests to ND site (which is very slow)
